Fixed Encoder Counter
Fixed Encoder Counter
