+++ src/org/usfirst/frc2783/commands/TankDrive.java
@@ -25,6 +25,25 @@ public class TankDrive extends Command {
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
     	
+    	if(OI.driver.getRawAxis(1) > 0.15){
+    		Robot.isLeftForward = true;
+    	}
+    	
+    	else if(OI.driver.getRawAxis(1) < -0.15){
+    		Robot.isLeftForward = false;
+    	}
+    	
+    	if(OI.driver.getRawAxis(5) > 0.15){
+    		Robot.isRightForward = true;
+    	}
+    	
+    	else if(OI.driver.getRawAxis(5) < -0.15){
+    		Robot.isRightForward = false;
+    	}
+    	
+    	lMot = OI.leftJoy.getRawAxis(1)/2;
+    	rMot = OI.rightJoy.getRawAxis(1)/2;
+    	
     	lMot = OI.driver.getRawAxis(1)/2;
     	rMot = OI.driver.getRawAxis(5)/2;
     	
@@ -37,15 +56,15 @@ public class TankDrive extends Command {
     		rMot = rMot*2;
     	}
    
-    	if(Math.abs(lMot) < 0.2){
+    	if(Math.abs(lMot) < 0.15){
     		lMot = 0;
     	}
     	
-    	if(Math.abs(rMot) < 0.2){
+    	if(Math.abs(rMot) < 0.15){
     		rMot = 0;
     	}
     	
-    	Robot.tankDriveBase.tankDrive(-lMot, rMot);
+    	Robot.tankDriveBase.tankDrive(lMot, rMot);
     	
     }
